split functionality between MP and TG
